Fix issue with size_t vs int Implement graphics to play pieces

+++ src/connect4_commandline.cpp
@@ -9,9 +9,9 @@
 namespace connect4 {
     void CommandLine::Display() const {
         std::string output;
-        for (size_t y = board.GetHeight(); y > 0; y--) {
+        for (int y = board.GetHeight(); y > 0; y--) {
             output += "|";
-            for (size_t x = 0; x < board.GetLength(); x++) {
+            for (int x = 0; x < board.GetLength(); x++) {
                 output += board.GetBoard().at(x).at(y - 1);
                 output += "|";
             }
